requirements.txt
alembic==1.8.1
anyio==3.6.1
asgiref==3.5.2
astroid==2.11.7
autopep8==1.6.0
bcrypt==3.2.2
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
dill==0.3.5.1
dnspython==2.2.1
ecdsa==0.18.0
email-validator==1.2.1
fastapi==0.79.0
greenlet==1.1.2
h11==0.13.0
httptools==0.4.0
idna==3.3
isort==5.10.1
itsdangerous==2.1.2
Jinja2==3.1.2
jose==1.0.0
lazy-object-proxy==1.7.1
Mako==1.2.1
MarkupSafe==2.1.1
mccabe==0.7.0
orjson==3.7.11
passlib==1.7.4
Pillow==9.2.0
platformdirs==2.5.2
psycopg2==2.9.3
pyasn1==0.4.8
pycodestyle==2.9.1
pycparser==2.21
pydantic==1.9.1
pylint==2.14.5
python-dotenv==0.20.0
python-jose==3.3.0
python-multipart==0.0.5
PyYAML==6.0
requests==2.28.1
rsa==4.9
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.39
starlette==0.19.1
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.4
typing_extensions==4.3.0
ujson==5.4.0
urllib3==1.26.11
uvicorn==0.17.6
uvloop==0.16.0
watchgod==0.8.2
websockets==10.3
wrapt==1.14.1